Make new searchbar actually work
